api: pool: split pool types into a separate header
Split pool types into a separate header file to enable easier function
inlining. No API changes.
