+++ include/odp/api/abi-default/pool.h
@@ -1,4 +1,4 @@
-/* Copyright (c) 2017-2018, Linaro Limited
+/* Copyright (c) 2022, Nokia
  * All rights reserved.
  *
  * SPDX-License-Identifier:     BSD-3-Clause
@@ -11,24 +11,7 @@
 extern "C" {
 #endif
 
-#include <odp/api/abi/event.h>
-
-/** @internal Dummy type for strong typing */
-typedef struct { char dummy; /**< @internal Dummy */ } _odp_abi_pool_t;
-
-/** @ingroup odp_pool
- *  @{
- */
-
-typedef _odp_abi_pool_t *odp_pool_t;
-
-#define ODP_POOL_INVALID   ((odp_pool_t)0)
-
-#define ODP_POOL_NAME_LEN  32
-
-/**
- * @}
- */
+/* Empty header required due to the packet inline functions */
 
 #ifdef __cplusplus
 }
