+++ include/odp/api/buffer.h
@@ -20,7 +20,6 @@ extern "C" {
 #include <odp/api/std_types.h>
 #include <odp/api/abi/event.h>
 #include <odp/api/abi/buffer.h>
-#include <odp/api/abi/pool.h>
 
 #include <odp/api/spec/buffer.h>
