+++ include/odp/api/classification.h
@@ -18,7 +18,6 @@ extern "C" {
 #endif
 
 #include <odp/api/std_types.h>
-#include <odp/api/abi/pool.h>
 #include <odp/api/abi/classification.h>
 #include <odp/api/abi/packet_types.h>
 #include <odp/api/abi/packet_io.h>
