+++ include/odp/api/crypto.h
@@ -21,7 +21,6 @@ extern "C" {
 #include <odp/api/abi/packet_types.h>
 #include <odp/api/abi/crypto.h>
 #include <odp/api/abi/buffer.h>
-#include <odp/api/abi/pool.h>
 #include <odp/api/abi/queue_types.h>
 
 /** @ingroup odp_crypto
