+++ include/odp/api/packet.h
@@ -23,7 +23,6 @@ extern "C" {
 #include <odp/api/abi/packet_types.h>
 #include <odp/api/abi/packet.h>
 #include <odp/api/abi/buffer.h>
-#include <odp/api/abi/pool.h>
 
 #include <odp/api/spec/packet.h>
