+++ include/odp/api/spec/dma.h
@@ -1,4 +1,4 @@
-/* Copyright (c) 2021, Nokia
+/* Copyright (c) 2021-2022, Nokia
  * All rights reserved.
  *
  * SPDX-License-Identifier:     BSD-3-Clause
@@ -19,6 +19,7 @@ extern "C" {
 #endif
 
 #include <odp/api/dma_types.h>
+#include <odp/api/pool_types.h>
 
 /** @addtogroup odp_dma
  *  @{
