+++ include/odp/api/spec/packet.h
@@ -22,6 +22,7 @@ extern "C" {
 #include <odp/api/proto_stats_types.h>
 #include <odp/api/time.h>
 #include <odp/api/packet_types.h>
+#include <odp/api/pool_types.h>
 
 /** @defgroup odp_packet ODP PACKET
  *  Packet event metadata and operations.
