+++ include/odp/api/spec/pool.h
@@ -1,5 +1,5 @@
 /* Copyright (c) 2015-2018, Linaro Limited
- * Copyright (c) 2020-2021, Nokia
+ * Copyright (c) 2020-2022, Nokia
  * All rights reserved.
  *
  * SPDX-License-Identifier:     BSD-3-Clause
@@ -20,7 +20,7 @@ extern "C" {
 #endif
 
 #include <odp/api/std_types.h>
-#include <odp/api/spec/pool_types.h>
+#include <odp/api/pool_types.h>
 
 /** @defgroup odp_pool ODP POOL
  *  Packet and buffer (event) pools.
