+++ include/odp/arch/x86_64-linux/odp/api/abi/pool_types.h
@@ -1,0 +1,7 @@
+/* Copyright (c) 2022, Nokia
+ * All rights reserved.
+ *
+ * SPDX-License-Identifier:     BSD-3-Clause
+ */
+
+#include <odp/api/abi-default/pool_types.h>
