+++ platform/linux-generic/include-abi/odp/api/abi/pool.h
@@ -1,4 +1,5 @@
 /* Copyright (c) 2015-2018, Linaro Limited
+ * Copyright (c) 2022, Nokia
  * All rights reserved.
  *
  * SPDX-License-Identifier:     BSD-3-Clause
@@ -17,23 +18,7 @@
 extern "C" {
 #endif
 
-#include <odp/api/std_types.h>
-#include <odp/api/plat/strong_types.h>
-#include <odp/api/abi/event.h>
-
-/** @ingroup odp_pool
- *  @{
- */
-
-typedef ODP_HANDLE_T(odp_pool_t);
-
-#define ODP_POOL_INVALID _odp_cast_scalar(odp_pool_t, 0)
-
-#define ODP_POOL_NAME_LEN  32
-
-/**
- * @}
- */
+/* Empty placeholder header for inline functions */
 
 #ifdef __cplusplus
 }
