+++ platform/linux-generic/odp_dma.c
@@ -1,4 +1,4 @@
-/* Copyright (c) 2021, Nokia
+/* Copyright (c) 2021-2022, Nokia
  * All rights reserved.
  *
  * SPDX-License-Identifier:     BSD-3-Clause
@@ -11,7 +11,10 @@
 #include <odp/api/buffer.h>
 #include <odp/api/stash.h>
 #include <odp/api/pool.h>
+#include <odp/api/queue.h>
+
 #include <odp/api/plat/std_inlines.h>
+
 #include <odp_global_data.h>
 #include <odp_debug_internal.h>
 #include <odp_init_internal.h>
